+++ spec/lib/human_error/errors/crud_errors/resource_persistence_error_spec.rb
@@ -1,20 +1,43 @@
 require 'rspectacular'
 require 'human_error/errors/crud_errors/resource_persistence_error'
+require 'active_support'
+require 'active_model'
+require 'active_record/errors'
+require 'active_record/validations'
+
+class HumanErrorTestModel
+  include ActiveModel::Model
+  include ActiveModel::AttributeMethods
+
+  attr_accessor :some_attribute
+
+  validates_presence_of :some_attribute
+
+  def attributes
+    {
+      'some_attribute' => @some_attribute,
+    }
+  end
+end
 
 class     HumanError
 module    Errors
 describe  ResourcePersistenceError do
-  let(:error) { ResourcePersistenceError.new }
-
   it 'has a status of 422' do
+    error = ResourcePersistenceError.new
+
     expect(error.http_status).to eql 422
   end
 
   it 'has a code of 1006' do
+    error = ResourcePersistenceError.new
+
     expect(error.code).to eql 1006
   end
 
   it 'has a knowledgebase article ID of 1234567890' do
+    error = ResourcePersistenceError.new
+
     expect(error.knowledgebase_article_id).to eql '1234567890'
   end
 
@@ -41,6 +64,28 @@ describe  ResourcePersistenceError do
 
     expect(error.friendly_message).to eql 'Sorry! We had a problem when tried to ' \
                                           'bullet time that black leather trenchcoat.'
+  end
+
+  it 'can convert an "ActiveRecord::RecordNotSaved"' do
+    record                     = HumanErrorTestModel.new
+    record.valid?
+    resource_persistence_error = ActiveRecord::RecordNotSaved.new('message', record)
+    error                      = ResourcePersistenceError.
+                                   convert(resource_persistence_error)
+
+    expect(error.attributes).to eql('some_attribute' => nil)
+    expect(error.errors).to     eql ["Some attribute can't be blank"]
+  end
+
+  it 'can convert an "ActiveRecord::RecordInvalid"' do
+    record                     = HumanErrorTestModel.new
+    record.valid?
+    resource_persistence_error = ActiveRecord::RecordInvalid.new(record)
+    error                      = ResourcePersistenceError.
+                                   convert(resource_persistence_error)
+
+    expect(error.attributes).to eql('some_attribute' => nil)
+    expect(error.errors).to     eql ["Some attribute can't be blank"]
   end
 end
 end